added exercises 4.15-4.23

+++ controllers/blogs.js
@@ -1,9 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   await Blog
     .find({})
+    .populate('user', { username: 1, name: 1 })
     .then(blogs => {
       response.json(blogs)
     })
@@ -18,20 +20,33 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  await blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })
-    .catch(error => {
-      response.status(400).json({ error: error.message })
-    })
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
+
+  const blog = new Blog({
+    ...body,
+    user: user.id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
+
 })
 
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+
+  const userId = request.user.id
+
+  if ( blog.user.toString() !== userId.toString()) {
+    return response.status(400).json({ error: 'user is not the owner of the blog' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
